+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stareezy-ui/core
 
+## 1.1.1
+
+### Patch Changes
+
+- add new docs & fix sx props logic
+- Updated dependencies
+  - @stareezy-ui/tokens@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes